+++ CHANGES.rst
@@ -23,6 +23,9 @@ There's a frood who really knows where his towel is.
 1.0b1 (2015-05-27)
 ------------------
 
+- Add control panel configlet to define global options for the Cycle2 slideshow plugin (closes `#3`_).
+  [hvelarde]
+
 - Remove browser layer and JS resource registry registrations.
   Include all Cycle2 and plugins code as browser resources (closes `#1`_).
   [rodfersou]
@@ -34,3 +37,4 @@ There's a frood who really knows where his towel is.
 - Initial release.
 
 .. _`#1`: https://github.com/collective/collective.js.cycle2/issues/1
+.. _`#3`: https://github.com/collective/collective.js.cycle2/issues/3